leetcode/src/test/java/_120TriangleTest.java
import org.junit.Assert;
import org.junit.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.LinkedList;
import java.util.List;

import static org.junit.Assert.*;

public class _120TriangleTest {

    @Test
    public void minimumTotal() {
        List<List<Integer>> triangle = new LinkedList<>();
        triangle.add(new ArrayList<>(Arrays.asList(2)));
        triangle.add(new ArrayList<>(Arrays.asList(3, 4)));
        triangle.add(new ArrayList<>(Arrays.asList(6, 5, 7)));
        triangle.add(new ArrayList<>(Arrays.asList(4, 1, 8, 3)));
        Assert.assertEquals(11, new _120Triangle().minimumTotal(triangle));
    }
}